+++ Offline_Scheduler/include/Napoleon/types.h
@@ -5,7 +5,7 @@
  * Project  : A DFG Off-Line Task Scheduler for FPGA
  * 
  * Created  : May 22, 2013
- * Modified : May 22, 2013
+ * Modified : May 30, 2013
  *************************************************************************/
 
 /*************************************************************************
@@ -31,14 +31,14 @@
 #define __ENQUEUE 1
 #define __DEQUEUE 2
 
-#define __ADD      1
-#define __MULT     2
-#define __SUB      3
-#define __DIV      4
-#define __LT       5
-#define __GT       6
-#define __LTE      7
-#define __GTE      8
+#define __ADD     1
+#define __MULT    2
+#define __SUB     3
+#define __DIV     4
+#define __LT      5
+#define __GT      6
+#define __LTE     7
+#define __GTE     8
 
 typedef struct s_config {
     char aif_fname[__SIZE_FILENAME];
@@ -49,48 +49,48 @@ typedef struct s_config {
 
 typedef struct s_task {
     char name[__SIZE_TASKNAME];
-    short int type;
-    short int exec_sched;
-    short int reconfig_sched;
-    short int leftmost_column;
-    short int bottommost_row;
-    short int latency;
-    short int columns;
-    short int rows;
-    short int reconfig_time;
-    short int width;
-    short int input1;
-    short int input2;
-    short int output;
-    short int impl;
-    short int reconfig_pwr;
-    short int exec_pwr;
+    int type;
+    int exec_sched;
+    int reconfig_sched;
+    int leftmost_column;
+    int bottommost_row;
+    int latency;
+    int columns;
+    int rows;
+    int reconfig_time;
+    int width;
+    int input1;
+    int input2;
+    int output;
+    int impl;
+    int reconfig_pwr;
+    int exec_pwr;
 } t_task;
 
 typedef struct s_task_interface {
     char name[__SIZE_TASK_INTFCNAME];
-    short int mode;
-    short int width;
-    short int reg_out;
+    int mode;
+    int width;
+    int reg_out;
 } t_task_interface;
 
 typedef struct s_task_Type {
-    short int latency;
-    short int reconfig_time;
-    short int columns;
-    short int rows;
+    int latency;
+    int reconfig_time;
+    int columns;
+    int rows;
 } t_task_type;
 
 typedef struct s_task_Impl {
-    short int id;
-    short int type;
-    short int impl;
-    short int latency;
-    short int reconfig_time;
-    short int columns;
-    short int rows;
-    short int reconfig_pwr;
-    short int exec_pwr;
+    int id;
+    int type;
+    int impl;
+    int latency;
+    int reconfig_time;
+    int columns;
+    int rows;
+    int reconfig_pwr;
+    int exec_pwr;
 } t_task_impl;
